pyproject.toml
[project]
name = "pacs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gdcm>=1.1",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "pydicom>=3.0.1",
    "pynetdicom>=2.1.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.1",
    "tenacity>=9.1.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
